# Improve common supertype calculation