Required for https://github.com/onflow/cadence/pull/2099

## Description

Improve the supertype calculation to make it easy to add new types in the 'upper mask' section.

This does not change the behaviour for the existing types. Therefore not a breaking change.

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [ ] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels