@turbolent 2022-10-28
Maybe add a benchmark which exercises `LeastCommonSuperType` and compare to before, then add a comment that gives a rough estimate for the overhead of the defensive check.