@SupunS 2022-10-28
Okayyy this sounds bad 
(well, this is a micro-benchmark and the times are in `ns`, so this wouldn't really be visible in the overall performance. But still..)
```
name                             old time/op    new time/op    delta
SuperTypeInference/integers-8      16.1ns ± 1%    65.7ns ± 1%  +308.53%  (p=0.016 n=4+5)
SuperTypeInference/arrays-8         164ns ±15%     225ns ± 2%   +36.98%  (p=0.008 n=5+5)
SuperTypeInference/composites-8    77.5ns ±12%    90.1ns ± 0%   +16.30%  (p=0.016 n=5+4)

name                             old alloc/op   new alloc/op   delta
SuperTypeInference/integers-8       0.00B          0.00B           ~     (all equal)
SuperTypeInference/arrays-8         96.0B ± 0%     96.0B ± 0%      ~     (all equal)
SuperTypeInference/composites-8     0.00B          0.00B           ~     (all equal)

name                             old allocs/op  new allocs/op  delta
SuperTypeInference/integers-8        0.00           0.00           ~     (all equal)
SuperTypeInference/arrays-8          3.00 ± 0%      3.00 ± 0%      ~     (all equal)
SuperTypeInference/composites-8      0.00           0.00           ~     (all equal)
```

Having the defensive check means if the type is inferred, the `IsSubType` check is run for the 'inferred' vs 'actual' types, just as if the user has declared the type (in that case 'expected' vs 'actual'). So that overhead would be there even if the user declared the type.

What's interesting is, inferring the type is faster than performing the subtype types check, for some cases, like integers 🤔
(test is here: [9994d8b](https://github.com/onflow/cadence/pull/2103/commits/9994d8ba1c02066d0a250b5a856e2e868f8e45c2))
```
SuperTypeInference/integers-8      16.1ns ± 1%    65.7ns ± 1%  +308.53% 
```
The overhead os subtype check is larger than the actual inference.

```cadence
let x = [1, 2, 3]      // type is inferred, so its fast
```

```cadence
let x: Int = [1, 2, 3]  // IsSubType runs, so its slow
```