@SupunS 2022-10-31
Offline meeting conclusion: The check happens at checking time, and also the overhead is in nano-seconds. So this doesn't really impact the type-checking performance visibly. At the same time, assuring this safety using a test case is also impossible, since there can be infinite number of combinations of types. So the defensive check is worth having.